# openapi.yaml
openapi: 3.1.0
info:
  title: DOGFOOD Portal API
  version: 1.0.0
  description: >-
    Local-first hackathon submission, judging, community, and organizer API.
    Session identity is carried in an HttpOnly session cookie. Runtime data is in memory.
servers:
  - url: http://localhost:8080
security: []
paths:
...
  /api/v1/teams:
    post:
      summary: Register a team and issue a participant session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email]
              properties:
                name: { type: string, maxLength: 80 }
                email: { type: string, format: email }
      responses:
        "201": { description: Team created; session cookie issued }
        "403": { description: Registration closed }
  /api/v1/team:
    get:
      summary: Read the signed-in participant's team
      security: [{ SessionCookie: [] }]
      responses:
        "200": { description: Team and members }
        "403": { description: Participant role required }
...
  /api/v1/teams/{id}/invitations:
    post:
      summary: Create an email-bound, expiring team invite
      security: [{ SessionCookie: [] }]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
      responses:
        "201": { description: One-time invite token }
        "403": { description: Team ownership required }
  /api/v1/invitations/{token}/accept:
    post:
      summary: Accept a team invite and issue a participant session
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200": { description: Membership accepted; session cookie issued }
        "404": { description: Invalid, expired, or used invite }
  /api/v1/projects:
    get:
      summary: Search public projects
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: track, in: query, schema: { type: string } }
      responses:
        "200": { description: Filtered project list }
    post:
      summary: Submit a project for the participant's team
      security: [{ SessionCookie: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectInput" }
      responses:
        "201": { description: Project created }
        "403": { description: Participant role or open deadline required }
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: session
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
  schemas:
...
    ProjectInput:
      type: object
      required: [title]
      properties:
        title: { type: string, maxLength: 120 }
        track: { type: string }
        summary: { type: string, maxLength: 2000 }
        repo_url: { type: string, format: uri }